extract usage-limit preflight from symptom-chat route
